+++ src/rads_toolbox/commands/format.py
@@ -1,12 +1,13 @@
+from subprocess import PIPE, CompletedProcess
+
 import click
-import invoke
 
 from rads_toolbox.contexts import AppContext, pass_app_context
-from rads_toolbox.utils import ensure_pre_commit, print_header, run_command_str
+from rads_toolbox.utils import OnError, print_header, run, run_command_str
 
 
-def _check_result(app_context: AppContext, result: invoke.Result, check: bool, msg: str):
-    if result.return_code == 1 and check:
+def _check_result(app_context: AppContext, result: CompletedProcess, check: bool, msg: str):
+    if result.returncode == 1 and check:
         click.secho(
             f"{msg} before commit. Try following:\n"
             f" * Enable pre-commit hook by running `pre-commit install` in the repository.\n"
@@ -16,7 +17,7 @@ def _check_result(app_context: AppContext, result: invoke.Result, check: bool, m
         )
         raise click.Abort()
 
-    if result.return_code > 1:
+    if result.returncode > 1:
         raise click.Abort()
 
 
@@ -26,10 +27,11 @@ def _check_result(app_context: AppContext, result: invoke.Result, check: bool, m
 @pass_app_context
 def run_format(app_context: AppContext, check: bool, quiet: bool):
     """Runs black code formatter and isort on source code."""
-    ensure_pre_commit(app_context.ctx)
+    # ensure pre-commit is installed
+    run("pre-commit install", stdout=PIPE, on_error=OnError.EXIT)
 
     toolbox = app_context.py_project_toml.tool.toolbox
-    dirs = f"{toolbox.sources_directory} {toolbox.tests_directory}"
+    dirs = [toolbox.sources_directory, toolbox.tests_directory]
     flags = []
 
     if check:
@@ -37,21 +39,11 @@ def run_format(app_context: AppContext, check: bool, quiet: bool):
 
     print_header("Sorting imports", icon="ℹ")
 
-    _check_result(
-        app_context,
-        app_context.ctx.run(f"isort {dirs} " + " ".join(flags), pty=True, warn=True),
-        check,
-        "Import were not sorted",
-    )
+    _check_result(app_context, run(["isort", *dirs, *flags], on_error=OnError.PASS), check, "Import were not sorted")
 
     print_header("Formatting code", icon="🖤")
 
     if quiet:
         flags.append("--quiet")
 
-    _check_result(
-        app_context,
-        app_context.ctx.run(f"black {dirs} " + " ".join(flags), pty=True, warn=True),
-        check,
-        "Code was not formatted",
-    )
+    _check_result(app_context, run(["black", *dirs, *flags], on_error=OnError.PASS), check, "Code was not formatted")